1. Extract define for second format in each view controller. 2. Modified the time type to float 3. Change button and label in table detail view controller.

+++ RecordAPP/ActionLevelToVCProtocol.h
@@ -10,7 +10,7 @@
 
 @protocol ActionLevelToVCProtocol <NSObject>
 
-- (int) getPerpareTime;
+- (float) getPerpareTime;
 #pragma mark (TODO) Change the name (not precise)
 - (float) getActionTime;
